+++ src/article_scanner/article_loader.py
@@ -1,7 +1,30 @@
+from sqlite3_interface import Database, DataTable, SQL_Query
 
-class article():
+def load_articles(db_conn: Database, load: str) -> DataTable:
 
-    def __init__(self):
-        self.title = self.content = None
-        
-        self.day = self.month = self.year = None
+    articles_table = DataTable(db_conn, "Articles")
+
+    if load == "hard":
+
+        sql_str = '''
+            CREATE TABLE IF NOT EXISTS Articles (
+                article_id INTEGER PRIMARY KEY,
+                article_title TEXT,
+                article_julian_day INTEGER,
+                article_content TEXT,
+                rep_id INTEGER,
+                FOREIGN KEY (rep_id)
+                REFERENCES Directory (rep_id)
+                    ON UPDATE CASCADE
+                    ON DELETE CASCADE
+            );
+        '''
+
+        articles_table.set_custom_query(sql_str).execute()
+
+    return articles_table
+
+def search_and_load_articles(db_conn, load) -> str:
+    '''returns 
+    a code of diagnostics like new articles added and 
+    success/failure codes'''
